convert labels to long

+++ train_cityscapes.py
@@ -271,7 +271,7 @@ def train(train_loader, model, criterion, optimizer, metric, epoch, args):
 
         # compute output
         output = model(input)
-        loss = criterion(output.view(output.shape[0], 19, -1), target.view(target.shape[0], -1))
+        loss = criterion(output.view(output.shape[0], 19, -1), target.view(target.shape[0], -1).long())
 
         # measure accuracy and record loss
         losses.update(loss.item(), input.size(0))
